PycharmProjects/basketball-franchise-manager/requirements.txt
passlib==1.7.2
flask==1.1.2
wheel==0.29.0
PyQt5==5.14.0
pandas==1.1.2
bs4==0.0.1
requests==2.24.0
pillow==7.2.0
MLB-StatsAPI
kivy==2.3.1
kivymd==1.2.0
kivymd_extensions.akivymd
buildozer
pyperclip
geopandas
